+++ src/utils.ts
@@ -14,6 +14,12 @@ export function delay(ms: number) {
     return new Promise(resolve => setTimeout(resolve, ms));
 }
 
+export function getRandomNumberBetween(min: number, max: number) {
+    min = Math.ceil(min);
+    max = Math.floor(max);
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
 export function shuffleArray<T>(array: T[]): T[] {
     let currentIndex = array.length, randomIndex;