+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Added Storm Spiral Animation macro - no sounds yet
  - Migrated Storm Spiral from Macro to Action triggered from template hook
+ - Added sound to Storm Spiral
 
 ## [0.52.0] - 2025-01-03